TADS_5 added menu and minor fixes

+++ TADS_3/inc/matrix_common.h
@@ -11,6 +11,6 @@ typedef long long mat_data_t;
 typedef unsigned short mat_index_t;
 
 // Кол-во итераций при подсчёте
-extern unsigned int mult_iterations;
+extern unsigned long int mult_iterations;
 
 #endif // _MATRIX_COMMON_H_
